Work done to update the system to a Raspberry Pi 4 running the latest version of Raspbian
The previous device (Raspberry Pi Zero) is insufficient for the data analysis and functions desired by the LESA team.

A faster single board computer is required, and this folder contains all the details, files, and relevant documentation to the work done to upgrade the project. This is still a work in progress.